Fix cyclic dependencies by moving Session to common (#414)
